CPM-257: Split attribute create button
